requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.1
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
tzdata==2024.2
urllib3==2.3.0
